+++ packages/gatsby/src/services/write-out-requires.ts
@@ -1,14 +1,14 @@
 import { IBuildContext } from "./"
 import reporter from "gatsby-cli/lib/reporter"
 import { writeAll } from "../bootstrap/requires-writer"
+import { assertStore } from "../utils/assert-store"
 
 export async function writeOutRequires({
   store,
   parentSpan,
 }: Partial<IBuildContext>): Promise<void> {
-  if (!store) {
-    reporter.panic(`No redux store`)
-  }
+  assertStore(store)
+
   // Write out files.
   const activity = reporter.activityTimer(`write out requires`, {
     parentSpan,