+++ www/src/components/markdown-page-footer.js
@@ -12,26 +12,17 @@ export default function MarkdownPageFooter(props) {
           sx={{
             display: `flex`,
             alignItems: `center`,
-            justifyContent: `space-between`,
             mt: 9,
           }}
         >
           <a
             sx={{ variant: `links.muted` }}
-            href={`https://github.com/gatsbyjs/gatsby/blob/master/${
-              props.packagePage ? `packages` : `docs`
-            }/${props.page ? props.page.parent.relativePath : ``}`}
+            href={`https://github.com/gatsbyjs/gatsby/blob/master/docs/${
+              props.page ? props.page.parent.relativePath : ``
+            }`}
           >
             <EditIcon sx={{ mr: 2 }} /> Edit this page on GitHub
           </a>
-          {props.page?.parent?.fields?.gitLogLatestDate && (
-            <span sx={{ color: `textMuted`, fontSize: 1 }}>
-              Last updated:{` `}
-              <time dateTime={props.page.parent.fields.gitLogLatestDate}>
-                {props.page.parent.fields.gitLogLatestDate}
-              </time>
-            </span>
-          )}
         </div>
       )}
     </Fragment>
@@ -43,19 +34,6 @@ export const fragment = graphql`
     parent {
       ... on File {
         relativePath
-        fields {
-          gitLogLatestDate(formatString: "MMMM D, YYYY")
-        }
-      }
-    }
-  }
-  fragment MarkdownPageFooter on MarkdownRemark {
-    parent {
-      ... on File {
-        relativePath
-        fields {
-          gitLogLatestDate(formatString: "MMMM D, YYYY")
-        }
       }
     }
   }