+++ packages/gatsby-plugin-page-creator/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-page-creator@2.3.14...gatsby-plugin-page-creator@2.3.15) (2020-07-02)
+
+**Note:** Version bump only for package gatsby-plugin-page-creator
+
+## [2.3.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-page-creator@2.3.13...gatsby-plugin-page-creator@2.3.14) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-plugin-page-creator
+
+## [2.3.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-page-creator@2.3.12...gatsby-plugin-page-creator@2.3.13) (2020-07-01)
+
+**Note:** Version bump only for package gatsby-plugin-page-creator
+
 ## [2.3.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-page-creator@2.3.11...gatsby-plugin-page-creator@2.3.12) (2020-06-24)
 
 **Note:** Version bump only for package gatsby-plugin-page-creator